feat: cc-0007 combined pricebar footer and pricebar
