+++ src/layouts/index.tsx
@@ -2,7 +2,6 @@ import { ReactNode } from "react";
 import { Box, Container } from "@chakra-ui/react";
 import Header from "./Header";
 import Footer from "./Footer";
-import PriceFooter from "../components/PriceFooter";
 import PriceBar from "../components/PriceBar";
 
 const Layout: React.FC<{ children: ReactNode }> = ({ children }) => {
@@ -14,8 +13,6 @@ const Layout: React.FC<{ children: ReactNode }> = ({ children }) => {
         <Footer />
       </Box>
       <PriceBar />
-      <PriceFooter />
-      
     </>
   );
 };
